Refactored and updated docs for panda
