+++ src/main/java/live/olszewski/bamboo/apiKeys/ApiKeysServiceImpl.java
@@ -3,9 +3,9 @@ package live.olszewski.bamboo.apiKeys;
 import live.olszewski.bamboo.apiResponse.ApiResponseBuilder;
 import live.olszewski.bamboo.apiResponse.ApiResponseDto;
 import live.olszewski.bamboo.auth.userStorage.UserStorage;
-import live.olszewski.bamboo.panda.PandaDao;
 import live.olszewski.bamboo.panda.PandaRepository;
 import live.olszewski.bamboo.panda.device.PandaService;
+import live.olszewski.bamboo.panda.objects.PandaDao;
 import live.olszewski.bamboo.panda.ownershipCheck.PandaOwnershipCkeck;
 import live.olszewski.bamboo.services.apiKey.ApiKeyService;
 import org.springframework.beans.factory.annotation.Autowired;
