+++ src/main/java/live/olszewski/bamboo/auth/ApiKeyFilter.java
@@ -24,6 +24,7 @@ public class ApiKeyFilter extends OncePerRequestFilter {
 
     private final ApiKeysService apiKeysService;
     private final PandaStorage pandaStorage;
+
     /**
      * Constructor for the ApiKeyFilter class.
      *
@@ -39,21 +40,22 @@ public class ApiKeyFilter extends OncePerRequestFilter {
      * It checks if the API key in the request header is valid.
      * If the API key is valid, it sets the authentication in the security context.
      * If the API key is not valid, it clears the security context.
-     * @param request The HTTP request.
-     * @param response The HTTP response.
+     *
+     * @param request     The HTTP request.
+     * @param response    The HTTP response.
      * @param filterChain The filter chain.
      * @throws ServletException If the request could not be handled.
-     * @throws IOException If an input or output error is detected when the servlet handles the request.
+     * @throws IOException      If an input or output error is detected when the servlet handles the request.
      */
     @Override
     protected void doFilterInternal(HttpServletRequest request, @NonNull HttpServletResponse response, @NonNull FilterChain filterChain) throws ServletException, IOException {
         String apiKey = request.getHeader("Authorization");
         if (apiKey != null && apiKey.startsWith("Bearer ")) {
             try {
-                System.out.println("API-KEY");
                 String apiKeyToken = apiKey.replace("Bearer ", "");
                 ApiKeyDto apiKeyDto = apiKeysService.verifyApiKey(apiKeyToken);
                 if (Objects.equals(apiKeyDto.getKey(), apiKeyToken)) {
+                    // If the API key is valid, set the authentication in the security context
                     pandaStorage.setCurrentPanda(apiKeyToken);
                     UsernamePasswordAuthenticationToken auth = new UsernamePasswordAuthenticationToken(apiKeyDto.getPanda(), null, null);
                     SecurityContextHolder.getContext().setAuthentication(auth);
