+++ src/main/java/live/olszewski/bamboo/panda/PandaRepository.java
@@ -1,5 +1,6 @@
 package live.olszewski.bamboo.panda;
 
+import live.olszewski.bamboo.panda.objects.PandaDao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
