+++ src/main/java/live/olszewski/bamboo/panda/config/PandaConfigController.java
@@ -11,6 +11,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+/**
+ * This controller handles requests related to Panda configuration.
+ */
 @RestController
 @RequestMapping(path = "#{apiConfig.pandaConfigPath}")
 @Tag(name = "Panda Configuration", description = "Endpoint to acquire panda configuration")
@@ -19,6 +22,12 @@ public class PandaConfigController {
     @Autowired
     private PandaConfigService pandaConfigService;
 
+    /**
+     * Retrieves the configuration for a specific Panda.
+     *
+     * @param id The ID of the Panda.
+     * @return A ResponseEntity containing the Panda's configuration.
+     */
     @GetMapping("/{id}")
     @Operation(
             description = "Get panda configuration",
@@ -49,6 +58,12 @@ public class PandaConfigController {
         return pandaConfigService.downloadPandaConfig(id);
     }
 
+    /**
+     * Verifies a Panda configuration.
+     *
+     * @param pandaConfigJson The Panda configuration in JSON format.
+     * @return A ResponseEntity containing the result of the verification.
+     */
     @GetMapping("/verify")
     @Operation(
             description = "Verify panda configuration",
