+++ src/main/java/live/olszewski/bamboo/panda/connection/PandaConnectionServiceImpl.java
@@ -23,41 +23,24 @@ public class PandaConnectionServiceImpl implements PandaConnectionService {
     @Autowired
     private ApiResponseBuilder apiResponseBuilder;
 
-    /**
-     * Checks the connection status of a Panda device using its UUID.
-     *
-     * @param UUID the UUID of the Panda device
-     * @return a PandaStatusDto object representing the connection status of the Panda device
-     * @throws IllegalStateException if no Panda device with the given UUID exists
-     */
+
     @Override
     public ResponseEntity<ApiResponseDto<?>> checkConnectionStatusWithUuid(String UUID) {
 
         Optional<PandaStatusDao> pandaStatusDao = pandaStatusRepository.findByUUID(UUID);
-        return pandaStatusDao.<ResponseEntity<ApiResponseDto<?>>>map(statusDao -> ResponseEntity.ok(apiResponseBuilder.buildSuccessResponse(200, "Check device connection status successfully", statusDao.toDto()))).orElseGet(() -> ResponseEntity.status(500).body(apiResponseBuilder.buildErrorResponse(500, "Panda with UUID: " + UUID + " does not exist")));
+        return pandaStatusDao.<ResponseEntity<ApiResponseDto<?>>>map(statusDao ->
+                        ResponseEntity.ok(apiResponseBuilder.buildSuccessResponse(200, "Check device connection status successfully", statusDao.toDto())))
+                .orElseGet(() -> ResponseEntity.status(500).body(apiResponseBuilder.buildErrorResponse(500, "Panda with UUID: " + UUID + " does not exist")));
     }
 
-    /**
-     * Checks the connection status of a Panda device using its ID.
-     *
-     * @param id the ID of the Panda device
-     * @return a PandaStatusDto object representing the connection status of the Panda device
-     * @throws IllegalStateException if no Panda device with the given ID exists
-     */
     @Override
     public ResponseEntity<ApiResponseDto<?>> checkConnectionStatusWithId(Long id) {
         Optional<PandaStatusDao> pandaStatusDao = pandaStatusRepository.findById(id);
-        if (pandaStatusDao.isEmpty())
-            return ResponseEntity.status(500).body(apiResponseBuilder.buildErrorResponse(500, "Panda with id: " + id.toString() + " does not exist"));
-        return ResponseEntity.ok(apiResponseBuilder.buildSuccessResponse(200, "Check device connection status successfully", pandaStatusDao.get().toDto()));
+        return pandaStatusDao.<ResponseEntity<ApiResponseDto<?>>>map(statusDao ->
+                        ResponseEntity.ok(apiResponseBuilder.buildSuccessResponse(200, "Check device connection status successfully", statusDao.toDto())))
+                .orElseGet(() -> ResponseEntity.status(500).body(apiResponseBuilder.buildErrorResponse(500, "Panda with id: " + id + " does not exist")));
     }
 
-    /**
-     * Sends the connection status of a Panda device to the repository.
-     *
-     * @param UUID   the UUID of the Panda device
-     * @param status the connection status of the Panda device
-     */
     @Override
     public ResponseEntity<ApiResponseDto<?>> sendConnectionStatus(String UUID, PandaStatus status) {
         PandaStatusDao pandaStatus = new PandaStatusDao(UUID, LocalDateTime.now(), status);
