+++ src/main/java/live/olszewski/bamboo/panda/dataPackets/collection/DataCollectionServiceImpl.java
@@ -3,10 +3,10 @@ package live.olszewski.bamboo.panda.dataPackets.collection;
 import live.olszewski.bamboo.apiResponse.ApiResponseBuilder;
 import live.olszewski.bamboo.apiResponse.ApiResponseDto;
 import live.olszewski.bamboo.auth.pandaStorage.PandaStorage;
-import live.olszewski.bamboo.panda.PandaDao;
 import live.olszewski.bamboo.panda.PandaRepository;
 import live.olszewski.bamboo.panda.dataPackets.DataPacketRepository;
 import live.olszewski.bamboo.panda.dataPackets.objects.DataPacketDao;
+import live.olszewski.bamboo.panda.objects.PandaDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -28,8 +28,7 @@ public class DataCollectionServiceImpl implements DataCollectionService {
 
     @Override
     public Long getPandaIdFromUUID(String uuid) {
-        PandaDao pandaDao = pandaRepository.findDeviceByUUID(uuid).orElse(null);
-        return pandaDao != null ? pandaDao.getId() : null;
+        return pandaRepository.findDeviceByUUID(uuid).map(PandaDao::getId).orElse(null);
     }
 
     @Override
