+++ src/main/java/live/olszewski/bamboo/panda/device/PandaController.java
@@ -10,6 +10,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+/**
+ * This controller handles requests related to Panda devices.
+ */
 @RestController
 @RequestMapping(path = "#{apiConfig.pandaDevicePath}")
 @Tag(name = "Panda", description = "Panda related endpoints")
@@ -18,6 +21,11 @@ public class PandaController {
     @Autowired
     private PandaService pandaService;
 
+    /**
+     * Retrieves all Panda devices.
+     *
+     * @return A ResponseEntity containing a list of all Panda devices.
+     */
     @GetMapping(path = "/all")
     @Operation(
             description = "Get all panda devices",
@@ -38,6 +46,11 @@ public class PandaController {
         return pandaService.getAllPandaDevices();
     }
 
+    /**
+     * Retrieves all Panda devices owned by the current user.
+     *
+     * @return A ResponseEntity containing a list of all Panda devices owned by the current user.
+     */
     @GetMapping()
     @Operation(
             description = "Get all panda devices by owner",
